feat: refactor zoom functionality and fix typos
- Replaced `onOpen` function with `onOpenZoom` in Item.vue
- Renamed ZoomViever.vue to ZoomViewer.vue for correct spelling
- Removed unused zoom.ts composable
- Removed ZoomModal from default layout
- Added new plain layout file
- Moved item page to a subdirectory and renamed it to index.vue
- Refactored zoom modal into a separate route under each item, renamed from ZoomModal.vue to zoom.vue
- Updated the way activeIndex is calculated in zoom view based on URL hash value
- Fetching item data directly in the zoom view instead of using global state.
